LayoutTests/fast/regions/cssom/element-region-overset-state-vertical-rl-expected.txt

Tests Element.webkitRegionOverset attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

  
PASS regionOverset('region_1') is "empty"
PASS regionOverset('region_2') is "empty"
PASS regionOverset('region_3') is "empty"
Flow content
PASS regionOverset('region_1') is "fit"
PASS regionOverset('region_2') is "fit"
PASS regionOverset('region_3') is "empty"
Add more content
PASS regionOverset('region_1') is "fit"
PASS regionOverset('region_2') is "fit"
PASS regionOverset('region_3') is "fit"
Add visual overflow
PASS regionOverset('region_1') is "fit"
PASS regionOverset('region_2') is "fit"
PASS regionOverset('region_3') is "fit"
Empty content
PASS regionOverset('region_1') is "empty"
PASS regionOverset('region_2') is "empty"
PASS regionOverset('region_3') is "empty"
PASS successfullyParsed is true

TEST COMPLETE
